Disable Github Flavored Markdown
GFM is causing issues with the line breaks, replacing every new line
with an HTML </br>, which causes the text to look ugly.

There is a solution to this problem (see
https://github.com/gettalong/kramdown/pull/83) but I could not get it to
work on my machine.

As GFM doesn't bring us significant advantages, and given that
@syrianspock and myself use standard Jekyll markdown for our respective
blogs it makes more sense to disable GFM for now.